requirements.txt
aiogram==3.27.0
aiohttp-socks==0.10.1
alembic==1.14.0
argon2-cffi==25.1.0
asyncpg==0.31.0
fastapi==0.115.12
itsdangerous==2.2.0
kryptoexpress-sdk==0.1.1
matplotlib==3.10.7
passlib==1.7.4
pillow==12.1.0
pydantic==2.9.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.21
qrcode==8.2
redis==6.0.0
sqladmin==0.22.0
SQLAlchemy==2.0.36
uvicorn==0.34.2
httpx==0.28.1
pyngrok

# Test dependencies
pytest==8.4.2
pytest-asyncio==1.2.0

python-decouple
